move events file
